+++ tests/test_sed.py
@@ -157,6 +157,110 @@ def test_SED_init():
     galsim.SED(spec=lambda w:1.0)
     galsim.SED(spec='1./(wave-700)')
 
+def test_SED_calculateMagnitude():
+    """ Check that magnitudes work as expected.
+    """
+    # Test that we can create a zeropoint with an SED, and that magnitudes for that SED are
+    # then 0.0
+    sed = galsim.SED(spec='wave')
+    bandpass = galsim.Bandpass(galsim.LookupTable([1,2,3,4,5], [1,2,3,4,5])).withZeropoint(sed)
+    np.testing.assert_almost_equal(sed.calculateMagnitude(bandpass), 0.0)
+    # Try multiplying SED by 100 to verify that magnitude decreases by 5
+    sed *= 100
+    np.testing.assert_almost_equal(sed.calculateMagnitude(bandpass), -5.0)
+    # Try setting zeropoint to a constant.
+    bandpass = galsim.Bandpass(galsim.LookupTable([1,2,3,4,5], [1,2,3,4,5])).withZeropoint(6.0)
+    np.testing.assert_almost_equal(sed.calculateMagnitude(bandpass),
+                                   (sed*100).calculateMagnitude(bandpass)+5.0)
+    # Try setting AB zeropoint
+    bandpass = (galsim.Bandpass(galsim.LookupTable([1,2,3,4,5], [1,2,3,4,5]))
+                .withZeropoint('AB', effective_diameter=640.0, exptime=15.0))
+    np.testing.assert_almost_equal(sed.calculateMagnitude(bandpass),
+                                   (sed*100).calculateMagnitude(bandpass)+5.0)
+
+    # See if we can set a magnitude.
+    sed = sed.withMagnitude(24.0, bandpass)
+    np.testing.assert_almost_equal(sed.calculateMagnitude(bandpass), 24.0)
+
+    # See if Vega magnitudes work.
+    # The following AB/Vega conversions are sourced from
+    # http://www.astronomy.ohio-state.edu/~martini/usefuldata.html
+    # Almost certainly, the LSST filters and the filters used on this website are not perfect
+    # matches, but should give some idea of the expected conversion between Vega magnitudes and AB
+    # magnitudes.  The results are consistent to 0.1 magnitudes, which is encouraging, but the true
+    # accuracy of the get/set magnitude algorithms is probably much better than this.
+    ugrizy_vega_ab_conversions = [0.91, -0.08, 0.16, 0.37, 0.54, 0.634]
+    filter_names = 'ugrizy'
+    for conversion, filter_name in zip(ugrizy_vega_ab_conversions, filter_names):
+        filter_filename = os.path.join(datapath, 'LSST_{0}.dat'.format(filter_name))
+        AB_bandpass = (galsim.Bandpass(filter_filename)
+                       .withZeropoint('AB', effective_diameter=640, exptime=15))
+        vega_bandpass = (galsim.Bandpass(filter_filename)
+                         .withZeropoint('vega', effective_diameter=640, exptime=15))
+        AB_mag = sed.calculateMagnitude(AB_bandpass)
+        vega_mag = sed.calculateMagnitude(vega_bandpass)
+        assert (abs((AB_mag - vega_mag) - conversion) < 0.1)
+
+    # Test intended meaning of zeropoint.  I.e., that an object with magnitude equal to the
+    # zeropoint will have a flux of 1.0.
+    bandpass = galsim.Bandpass(galsim.LookupTable([1,2,3,4,5], [1,2,3,4,5])).withZeropoint(24.0)
+    sed = sed.withMagnitude(bandpass.zeropoint, bandpass)
+    np.testing.assert_almost_equal(sed.calculateFlux(bandpass), 1.0, 10)
+
+def test_SED_calculateDCRMomentShifts():
+    # compute some moment shifts
+    sed = galsim.SED(os.path.join(datapath, 'CWW_E_ext.sed'))
+    bandpass = galsim.Bandpass(os.path.join(datapath, 'LSST_r.dat'))
+    Rbar, V = sed.calculateDCRMomentShifts(bandpass, zenith_angle=45*galsim.degrees)
+    # now rotate parallactic angle 180 degrees, and see if the output makes sense.
+    Rbar2, V2 = sed.calculateDCRMomentShifts(bandpass, zenith_angle=45*galsim.degrees,
+                                             parallactic_angle=180*galsim.degrees)
+    np.testing.assert_array_almost_equal(Rbar, -Rbar2, 15)
+    np.testing.assert_array_almost_equal(V, V2, 25)
+    # now rotate parallactic angle 90 degrees.
+    Rbar3, V3 = sed.calculateDCRMomentShifts(bandpass, zenith_angle=45*galsim.degrees,
+                                             parallactic_angle=90*galsim.degrees)
+    np.testing.assert_almost_equal(Rbar[0], Rbar3[1], 15)
+    np.testing.assert_almost_equal(V[1,1], V3[0,0], 25)
+    # and now test against an external known result.
+    np.testing.assert_almost_equal(V[1,1] * (180.0/np.pi * 3600)**2, 0.0065, 4)
+
+def test_SED_calculateSeeingMomentRatio():
+    # compute a relative moment shift and compare to externally generated known result.
+    sed = galsim.SED(os.path.join(datapath, 'CWW_E_ext.sed'))
+    bandpass = galsim.Bandpass(os.path.join(datapath, 'LSST_r.dat'))
+    relative_size = sed.calculateSeeingMomentRatio(bandpass)
+    np.testing.assert_almost_equal(relative_size, 0.919577157172, 4)
+
+def test_fnu_vs_flambda():
+    c = 2.99792458e17  # speed of light in nm/s
+    h = 6.62606957e-27 # Planck's constant in erg seconds
+    k = 1.3806488e-16  # Boltzmann's constant ergs per Kelvin
+    nm_in_cm = 1e7
+    # read these straight from Wikipedia
+    def rayleigh_jeans_fnu(T, w):
+        nu = c/w
+        return 2*nu**2*k*T/c**2 * nm_in_cm**2 # should have units of erg/s/cm^2/Hz
+    def rayleigh_jeans_flambda(T, w):
+        return 2*c*k*T / w**4 * nm_in_cm**2 # should have units of erg/s/cm^2/nm
+    waves = np.linspace(500, 1000, 100)
+    fnu = rayleigh_jeans_fnu(5800, waves)
+    flambda = rayleigh_jeans_flambda(5800, waves)
+
+    sed1 = galsim.SED(galsim.LookupTable(waves, fnu), flux_type='fnu')
+    sed2 = galsim.SED(galsim.LookupTable(waves, flambda), flux_type='flambda')
+    np.testing.assert_array_almost_equal(sed1(waves)/sed2(waves), np.ones(len(waves)), 10,
+                                         err_msg="Check fnu & flambda consistency.")
+
+    # Now also check that wavelengths in Angstroms work.
+    waves_ang = waves * 10
+    sed3 = galsim.SED(galsim.LookupTable(waves_ang, fnu), flux_type='fnu', wave_type='Ang')
+    sed4 = galsim.SED(galsim.LookupTable(waves_ang, flambda), flux_type='flambda', wave_type='Ang')
+    np.testing.assert_array_almost_equal(sed1(waves)/sed3(waves), np.ones(len(waves)), 10,
+                                         err_msg="Check nm and Ang SED wavelengths consistency.")
+    np.testing.assert_array_almost_equal(sed2(waves)/sed4(waves), np.ones(len(waves)), 10,
+                                         err_msg="Check nm and Ang SED wavelengths consistency.")
+
 if __name__ == "__main__":
     test_SED_add()
     test_SED_sub()
@@ -165,3 +269,7 @@ if __name__ == "__main__":
     test_SED_atRedshift()
     test_SED_roundoff_guard()
     test_SED_init()
+    test_SED_calculateMagnitude()
+    test_SED_calculateDCRMomentShifts()
+    test_SED_calculateSeeingMomentRatio()
+    test_fnu_vs_flambda()