+++ CHANGELOG.md
@@ -4,15 +4,18 @@ Changes from v1.1 to v1.2
 New Features
 ------------
 
-* Added an option to impose N-fold symmetry (for user-selected even values of
+- Added magnitudes as a method to set the flux of SED objects. (#547)
+- Added SED.calculateDCRMomentShifts and SED.calculateChromaticSeeingRatio convenience functions
+  for estimating chromatic PSF moment shifts (#547)
+- Added an option to impose N-fold symmetry (for user-selected even values of
   N>=4) on correlated noise fields, parallel to the existing noise whitening
   options. (#529)
 
 Bug Fixes and Improvements
 --------------------------
 
+- SED and Bandpass methods that return a new SED or Bandpass now attempt to preserve the type of
+  the object calling the method. (#547)
 
 Updates to config options
 -------------------------
-
-